+++ CHANGELOG.md
@@ -12,6 +12,16 @@ the following form: YYYY.0M.0D.
 - Fixed broken links
 - Disabled AI features globally
 
+## [2026.04.22]
+
+### Added
+- Added `BMSTest.java` class and 'BMS1Test.java' to test kernel and secondary methods.
+- Added 2 use cases 'VehicleDashboard.java' and 'BatteryPackRepair.java'.
+- Added the discussion of the project.
+
+### Updated
+- Updated file structure.
+
 ## [2026.04.14]
 
 ### Added